requirements.txt
fastapi
uvicorn
flake8
pre-commit
pydantic
sqlalchemy
sqlmodel
sqlalchemy[asyncio]
alembic
asyncpg
python-dotenv
alembic
